requirements.txt
numpy==1.23.5
python-dotenv==0.21.1
pusher==3.3.1
